refactor(NodeActivator): split module
