+++ utilities/node_activator/test/node_activator_test.exs
@@ -2,6 +2,9 @@ defmodule NodeActivatorTest do
   use ExUnit.Case
   doctest NodeActivator
 
+  alias NodeActivator.Epmd
+  alias NodeActivator.Utils
+
   setup do
     on_exit(fn -> Node.stop() end)
   end
@@ -25,41 +28,9 @@ defmodule NodeActivatorTest do
     end
   end
 
-  describe "generate_node_name" do
-    test "generates unique atom" do
-      node_name1 = NodeActivator.generate_node_name("test prefix")
-      node_name2 = NodeActivator.generate_node_name("test prefix")
-
-      assert is_atom(node_name1)
-      refute node_name1 == node_name2
-    end
-
-    test "starts with provided prefix" do
-      node_name = NodeActivator.generate_node_name("test prefix")
-
-      assert node_name |> Atom.to_string() |> String.starts_with?("test-prefix_")
-    end
-
-    test "ends with hostname" do
-      node_name = NodeActivator.generate_node_name("test prefix")
-      hostname = NodeActivator.get_hostname()
-
-      assert node_name |> Atom.to_string() |> String.ends_with?(hostname)
-    end
-  end
-
-  describe "get_hostname" do
-    test "returns non-blank string" do
-      hostname = NodeActivator.get_hostname()
-
-      assert is_binary(hostname)
-      assert String.length(hostname) > 1
-    end
-  end
-
   describe "launch_epmd" do
     test "does not raise error" do
-      assert :ok = NodeActivator.launch_epmd()
+      assert :ok = Epmd.launch_epmd()
     end
   end
 end
